ndo2db: Object cache & hash improvements.
Change NULL name2 to empty for the object cache to simplify and
optimize lookup and comparison. Empty name2 is still converted to NULL
for the DB.

The object hash table is only used by the object cache so keep it
internal to that implementation. Resize the object hash table based on
the number of objects in ndo2db_load_obj_cache() to reduce collisions
for large configurations. Count objects added and collisions.

Cache objects loaded from the DB when not originally found in the
cache so we don't keep going to the DB for later lookups. This
shouldn't happen since we load all objects from the DB on startup, but
it's possible for ndo2db_load_obj_cache() to not successfully process
the complete result set.

Remove original hash algorithm/constants.